The application side has no limit validation for adding 3 devices only. It is a rare case scenario that end-users will be adding multiple devices more than 3 but still Ochsner team would not like to restrict anything from the application code perspective.
The application side has no restriction validation for adding 2 rescues and one maintenance.

The Device screen should be medicine-driven and not device-driven. There can be scenarios where single medicine would be used for both rescue and maintenance as well.

If a user has already previously paired devices and has data in the database, the user should be re-directed on Dashboard based on the API device list